+++ worker/src/RTC/DataConsumer.cpp
@@ -152,7 +152,7 @@ namespace RTC
 		return FBS::DataConsumer::CreateGetStatsResponseDirect(
 		  builder,
 		  // timestamp.
-		  this->shared->GetTimeMs(),
+		  static_cast<uint64_t>(this->shared->GetTimeMs()),
 		  // label.
 		  this->label.c_str(),
 		  // protocol.
